models/project/requirements.txt
absl-py==0.8.1
asn1crypto==0.24.0
astor==0.8.0
cryptography==2.1.4
cycler==0.10.0
# Editable install with no version control (fashion-mnist==0.1)
-e /app
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.3
h5py==2.10.0
idna==2.6
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.2
numpy==1.17.3
opencv-python==4.1.2.30
opt-einsum==3.1.0
pandas==1.0.0
protobuf==3.10.0
pycrypto==2.6.1
pygobject==3.26.1
pyparsing==2.4.6
python-apt==1.6.4
python-dateutil==2.8.1
pytz==2019.3
pyxdg==0.25
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
SecretStorage==2.3.1
six==1.11.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
Werkzeug==0.16.0
wrapt==1.11.2
